Clean rename.
