+++ src/components/Editor/BacklinkSuggestionsDisplay.tsx
@@ -3,7 +3,7 @@ import React, { useRef, useEffect, useState, useMemo } from 'react'
 
 import posthog from 'posthog-js'
 
-import { removeFileExtension } from '@/lib/strings'
+import { removeFileExtension } from '@/lib/file'
 
 export interface SuggestionsState {
   textWithinBrackets: string
