+++ src/components/File/RenameDirectory.tsx
@@ -5,7 +5,7 @@ import { toast } from 'react-toastify'
 
 import ReorModal from '../Common/Modal'
 
-import { getInvalidCharacterInFileName } from '@/lib/strings'
+import { getInvalidCharacterInFileName } from '@/lib/file'
 import { useFileContext } from '@/contexts/FileContext'
 
 const RenameDirModal: React.FC = () => {
