+++ src/components/File/RenameNote.tsx
@@ -1,36 +1,21 @@
-import React, { useEffect, useState } from 'react'
+import React, { useState } from 'react'
 
 import { Button } from '@material-tailwind/react'
 import { toast } from 'react-toastify'
 
 import ReorModal from '../Common/Modal'
 
-import { getInvalidCharacterInFileName, removeFileExtension } from '@/lib/strings'
+import { getInvalidCharacterInFileName } from '@/lib/file'
 import { useFileContext } from '@/contexts/FileContext'
 
 const RenameNoteModal: React.FC = () => {
   const { noteToBeRenamed, setNoteToBeRenamed, renameFile } = useFileContext()
-
-  const fileExtension = noteToBeRenamed.split('.').pop() || 'md'
-  const [dirPrefix, setDirPrefix] = useState<string>('')
-  const [noteName, setNoteName] = useState<string>('')
+  const [newNoteName, setNewNoteName] = useState<string>('')
   const [errorMessage, setErrorMessage] = useState<string | null>(null)
-
-  useEffect(() => {
-    const setDirectoryUponNoteChange = async () => {
-      const initialNotePathPrefix = await window.path.dirname(noteToBeRenamed)
-      setDirPrefix(initialNotePathPrefix)
-      const initialNoteName = await window.path.basename(noteToBeRenamed)
-      const trimmedInitialNoteName = removeFileExtension(initialNoteName) || ''
-      setNoteName(trimmedInitialNoteName)
-    }
-
-    setDirectoryUponNoteChange()
-  }, [noteToBeRenamed])
 
   const handleNameChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const newName = e.target.value
-    setNoteName(newName)
+    setNewNoteName(newName)
 
     getInvalidCharacterInFileName(newName).then((invalidCharacter) => {
       if (invalidCharacter) {
@@ -44,11 +29,11 @@ const RenameNoteModal: React.FC = () => {
     setNoteToBeRenamed('')
   }
 
-  const sendNoteRename = async () => {
+  const handleNoteRename = async () => {
     if (errorMessage) {
       return
     }
-    if (!noteName) {
+    if (!newNoteName) {
       toast.error('Note name cannot be empty', {
         className: 'mt-5',
         closeOnClick: false,
@@ -57,13 +42,17 @@ const RenameNoteModal: React.FC = () => {
       return
     }
 
-    await renameFile(noteToBeRenamed, `${dirPrefix}${noteName}.${fileExtension}`)
+    const initialNotePathPrefix = await window.path.dirname(noteToBeRenamed)
+    const initialPath = await window.path.join(initialNotePathPrefix, newNoteName)
+    const outputPath = await window.path.addExtensionIfNoExtensionPresent(initialPath)
+
+    await renameFile(noteToBeRenamed, outputPath)
     onClose()
   }
 
   const handleKeyPress = (e: React.KeyboardEvent<HTMLInputElement>) => {
     if (e.key === 'Enter') {
-      sendNoteRename()
+      handleNoteRename()
     }
   }
 
@@ -74,7 +63,7 @@ const RenameNoteModal: React.FC = () => {
         <input
           type="text"
           className=" block w-full rounded-md border border-gray-300 px-3 py-2 transition duration-150 ease-in-out focus:border-blue-300 focus:outline-none"
-          value={noteName}
+          value={newNoteName}
           onChange={handleNameChange}
           onKeyDown={handleKeyPress}
           placeholder="New Note Name"
@@ -82,7 +71,7 @@ const RenameNoteModal: React.FC = () => {
         <div className="flex items-center gap-3">
           <Button
             className="mb-2 mt-3 h-10 w-[80px] cursor-pointer border-none bg-blue-500 px-2 py-0 text-center hover:bg-blue-600"
-            onClick={sendNoteRename}
+            onClick={handleNoteRename}
             placeholder=""
           >
             Rename
